requirements.txt
streamlit
numpy
matplotlib
Pillow
soundfile
openai
